test(T08/T10): migrate test_sql_exec to _helpers, remove weak until test
T08: add optional `rowcount` param and `_make_no_result_conn` to the
shared `_helpers` module, then remove the duplicate local definitions
from `test_sql_exec.py` and import from `_helpers` instead.  All six
SQL-service test modules now use the canonical factories.

T10: remove `test_list_request_history_until_adds_where_clause` — it
only asserted `"WHERE" in call_sql` (weaker than the `since` variant)
and is fully superseded by the binding-level test added in the same PR.

T02 nit: expand the pool-discard comment with an explicit TODO so
future readers understand the assertion documents a *gap*, not an
invariant.
