+++ tests/unit/services/test_query_insights.py
@@ -150,17 +150,6 @@ async def test_list_request_history_since_adds_where_clause() -> None:
     assert "submit_time" in call_sql
 
 
-async def test_list_request_history_until_adds_where_clause() -> None:
-    target = _make_target()
-    conn = _make_conn([], _REQ_HIST_COLS)
-    until = datetime(2024, 12, 31, tzinfo=UTC)
-    with patch("fabric_dw.sql.open_connection", return_value=conn):
-        await query_insights.list_request_history(target, until=until)
-    cursor = conn.cursor.return_value
-    call_sql: str = cursor.execute.call_args[0][0]
-    assert "WHERE" in call_sql
-
-
 async def test_list_request_history_maps_permission_denied() -> None:
     target = _make_target()
     conn = MagicMock()
